Backend/requirements.txt
# Core FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Production Server
gunicorn>=21.2.0


# Database
sqlmodel>=0.0.22
pg8000>=1.30.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.1
python-multipart>=0.0.12

# Machine Learning for Anomaly Detection
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
optuna>=3.0.0
apscheduler>=3.10.0

